Added Copyright, moved the mappings class and added Config class for later.
