+++ src/main/java/org/granitemc/granite/api/GraniteAPI.java
@@ -1,13 +1,33 @@
-/**
- * This file was created as part of Granite. <br>
- * @author matheus
- */
 package org.granitemc.granite.api;
 
 import org.granitemc.granite.events.EventBus;
 
 import java.util.ArrayList;
 
+/**
+ * License (MIT)
+ *
+ * Copyright (c) 2014. avarisc
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this
+ * software and associated documentation files (the "Software"), to deal in the
+ * Software without restriction, including without limitation the rights to use, copy,
+ * modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
+ * and to permit persons to whom the Software is furnished to do so, subject to the
+ * following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+ * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+@SuppressWarnings("ReflectionForUnavailableAnnotation")
 public class GraniteAPI {
 
     private static GraniteAPI instance;
@@ -18,6 +38,13 @@ public class GraniteAPI {
             instance = new GraniteAPI();
         }
         return instance;
+    }
+
+    public static EventBus eventBus() {
+        if (eventbus == null) {
+            eventbus = new EventBus(new ArrayList<Class<?>>());
+        }
+        return eventbus;
     }
 
     public boolean isPlugin(Object obj) {
@@ -44,13 +71,6 @@ public class GraniteAPI {
         PluginContainer ret = new PluginContainer();
         ret.loadFromPlugin(clazz);
         return ret;
-    }
-
-    public static EventBus eventBus() {
-        if (eventbus == null) {
-            eventbus = new EventBus(new ArrayList<Class<?>>());
-        }
-        return eventbus;
     }
 
 }
