+++ src/main/java/org/granitemc/granite/api/PluginContainer.java
@@ -10,6 +10,30 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
+/**
+ * License (MIT)
+ *
+ * Copyright (c) 2014. avarisc
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this
+ * software and associated documentation files (the "Software"), to deal in the
+ * Software without restriction, including without limitation the rights to use, copy,
+ * modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
+ * and to permit persons to whom the Software is furnished to do so, subject to the
+ * following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+ * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+@SuppressWarnings("ReflectionForUnavailableAnnotation")
 public class PluginContainer {
 
     private String name;
@@ -28,10 +52,6 @@ public class PluginContainer {
         setPluginClass(pluginClass);
     }
 
-    private void setName(String name2) {
-        name = name2;
-    }
-
     private void setID(String id2) {
         id = id2;
     }
@@ -48,6 +68,10 @@ public class PluginContainer {
      */
     public String getName() {
         return name;
+    }
+
+    private void setName(String name2) {
+        name = name2;
     }
 
     /**
