+++ src/main/java/org/granitemc/granite/events/EventBus.java
@@ -9,6 +9,29 @@ import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * License (MIT)
+ *
+ * Copyright (c) 2014. avarisc
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this
+ * software and associated documentation files (the "Software"), to deal in the
+ * Software without restriction, including without limitation the rights to use, copy,
+ * modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
+ * and to permit persons to whom the Software is furnished to do so, subject to the
+ * following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+ * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
 public class EventBus {
 
     /**
@@ -16,14 +39,19 @@ public class EventBus {
      */
     private boolean silent;
 
-    @Retention(RetentionPolicy.RUNTIME)
-    @Target(ElementType.METHOD)
-    public @interface EventSubscribed {}
-
     /**
      * All the handlers
      */
     private List<Class<?>> handlers;
+
+    /**
+     * Creates a new event bus with the given list of handlers
+     *
+     * @param handlers the handlers list
+     */
+    public EventBus(List<Class<?>> handlers) {
+        this.handlers = handlers;
+    }
 
     /**
      * Posts an event
@@ -66,8 +94,7 @@ public class EventBus {
     public void register(Class<?> eventHandler) {
         if (handlers != null) {
             handlers.add(eventHandler);
-        }
-        else {
+        } else {
             throw new IllegalArgumentException("This EventBus was not initialized.");
         }
     }
@@ -85,12 +112,8 @@ public class EventBus {
         }
     }
 
-    /**
-     * Creates a new event bus with the given list of handlers
-     *
-     * @param handlers the handlers list
-     */
-    public EventBus(List<Class<?>> handlers) {
-        this.handlers = handlers;
+    @Retention(RetentionPolicy.RUNTIME)
+    @Target(ElementType.METHOD)
+    public @interface EventSubscribed {
     }
 }
