+++ src/main/java/org/granitemc/granite/item/ItemStack.java
@@ -1,8 +1,31 @@
 package org.granitemc.granite.item;
 
-import org.granitemc.granite.Mappings;
+import org.granitemc.granite.utils.Mappings;
 
 import java.lang.reflect.InvocationTargetException;
+
+/**
+ * License (MIT)
+ *
+ * Copyright (c) 2014. avarisc
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this
+ * software and associated documentation files (the "Software"), to deal in the
+ * Software without restriction, including without limitation the rights to use, copy,
+ * modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
+ * and to permit persons to whom the Software is furnished to do so, subject to the
+ * following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+ * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
 
 public class ItemStack {
 
@@ -23,12 +46,12 @@ public class ItemStack {
         return (int) invoke("getItemDamage");
     }
 
-    public String[] getItemLore() {
-        return (String[]) invoke("getItemLore");
-    }
-
     public void setItemDamage(int damage) {
         invoke("setItemDamage", damage);
+    }
+
+    public String[] getItemLore() {
+        return (String[]) invoke("getItemLore");
     }
 
     public int getMaxDamage() {
