+++ src/main/java/org/granitemc/granite/utils/Mappings.java
@@ -1,8 +1,7 @@
-package org.granitemc.granite;
+package org.granitemc.granite.utils;
 
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
-import org.granitemc.granite.utils.Logger;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
@@ -11,25 +10,30 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 
+/**
+ * License (MIT)
+ *
+ * Copyright (c) 2014. avarisc
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this
+ * software and associated documentation files (the "Software"), to deal in the
+ * Software without restriction, including without limitation the rights to use, copy,
+ * modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
+ * and to permit persons to whom the Software is furnished to do so, subject to the
+ * following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+ * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
 public class Mappings {
-
-    public static class MappingNotFoundException extends RuntimeException {
-
-        /**
-         * So that eclipse stops bitching.
-         */
-        private static final long serialVersionUID = 1L;
-        private String triedToAccess;
-
-        public MappingNotFoundException(String triedToAccess) {
-            super("Tried to access " + triedToAccess);
-            this.triedToAccess = triedToAccess;
-        }
-
-        public String getTriedToAccess() {
-            return triedToAccess;
-        }
-    }
 
     private static Map<String, Class<?>> classes;
     private static Map<String, Map<String, ?>> methods;
@@ -132,6 +136,20 @@ public class Mappings {
         } catch (IllegalAccessException | InvocationTargetException e) {
             e.printStackTrace();
             throw new RuntimeException(e.getCause());
+        }
+    }
+
+    public static class MappingNotFoundException extends RuntimeException {
+
+        private String triedToAccess;
+
+        public MappingNotFoundException(String triedToAccess) {
+            super("Tried to access " + triedToAccess);
+            this.triedToAccess = triedToAccess;
+        }
+
+        public String getTriedToAccess() {
+            return triedToAccess;
         }
     }
 }
